+++ frontend/packages/core/src/Layout/index.tsx
@@ -1,0 +1,3 @@
+export { default as Grid } from "./grid";
+
+export type { GridJustification } from "./grid";
