+++ frontend/packages/core/src/icon.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import FiberManualRecordTwoToneIcon from "@mui/icons-material/FiberManualRecordTwoTone";
-import { Grid } from "@mui/material";
 
-import type { GridJustification } from "./grid";
+import type { GridJustification } from "./Layout";
+import { Grid } from "./Layout";
 import { styled } from "./Utils";
 
 const StyledStatusIcon = styled(FiberManualRecordTwoToneIcon)`
