+++ frontend/packages/core/src/index.tsx
@@ -17,7 +17,6 @@ export { useWorkflowStorageContext, useWizardContext, WizardContext } from "./Co
 export { Dialog, DialogActions, DialogContent } from "./dialog";
 export * from "./Feedback";
 export { FeatureOff, FeatureOn, SimpleFeatureFlag } from "./flags";
-export { default as Grid } from "./grid";
 export { AvatarIcon, StatusIcon } from "./icon";
 export * from "./Input";
 export { Link } from "./link";
@@ -51,11 +50,12 @@ export { default as Code } from "./text";
 export { default as TimeAgo } from "./timeago";
 export { Typography } from "./typography";
 export { default as ClutchApp } from "./AppProvider";
+export * from "./Layout";
 
 export type { BaseWorkflowProps, WorkflowConfiguration } from "./AppProvider/workflow";
 export type { ButtonProps } from "./button";
 export type { CardHeaderSummaryProps } from "./card";
-export type { GridJustification } from "./grid";
+
 export type { WorkflowRemoveDataFn, WorkflowRetrieveDataFn, WorkflowStoreDataFn } from "./Contexts";
 export type { ClutchError } from "./Network/errors";
 export type { TypographyProps } from "./typography";
